requirement.txt
kafka-python==2.0.2
requests==2.31.0
pyspark==3.5.1
streamlit==1.38.0
streamlit-autorefresh==1.0.1
plotly==5.22.0
pandas==2.2.2

